refactor: correct cmake file glob, use size_t instead of int everywhere

+++ src/heap/heap.h
@@ -11,13 +11,13 @@ struct Heap {
     void push(Value value);
     std::optional<Value> pop();
 
-    void debug(int start = 0);
+    void debug(size_t start = 0);
     const std::vector<Value>& values();
 
    private:
     Compare compare;
-    void bubble(int index);
-    void sink(int index);
+    void bubble(size_t index);
+    void sink(size_t index);
     std::vector<Value> _values;
 };
